Deque

Description

Double-ended queue (deque) purely functional implementation based on Chris Okasaki's "Purely functional data structures"

Documentation

data Deque a

Data type for deques

emptyDEQ :: Deque a

Creates an empty deque

O(1)

isEmptyDEQ :: Deque a -> Bool

Returns true if deque is empty

O(1)

lengthDEQ :: Deque a -> Int

Returns length of deque

O(1)

firstDEQ :: Deque a -> Maybe a

Returns item from the front of deque, without removing it

O(1)

lastDEQ :: Deque a -> Maybe a

Returns item from the back of deque, without removing it

O(1)

takeFrontDEQ :: Int -> Deque a -> [a]

Returns first n elements of deque

O(n)

takeBackDEQ :: Int -> Deque a -> [a]

Returns last n elements of deque

O(n)

pushFrontDEQ :: Deque a -> a -> Deque a

Pushes element to the front of deque and returns updated deque

O(1) amortised

popFrontDEQ :: Deque a -> Maybe (a, Deque a)

Pops element from the front of deque and returns it alongside updated deque

O(1) amortised

pushBackDEQ :: Deque a -> a -> Deque a

Pushes element to the back of deque and returns updated deque

O(1) amortised

popBackDEQ :: Deque a -> Maybe (a, Deque a)

Pops element from the back of deque and returns it alongside updated deque

O(1) amortised

reverseDEQ :: Deque a -> Deque a

Returns reversed deque

O(1)

fromListDEQ :: [a] -> Deque a

Creates deque from a list

O(n)

toListDEQ :: Deque a -> [a]

Creates list from a deque

O(n)

getDeque :: Maybe (a, Deque a) -> Deque a

Gets deque from result of either pop function

O(1)

isBalanced :: Deque a -> Bool

Returns true if deque is balanced

O(1)