Deque.Lazy

Description

Definitions of lazy Deque.

The typical toList and fromList conversions are provided by means of the Foldable and IsList instances.

Documentation

data Deque a

Lazy double-ended queue (aka Dequeue or Deque) based on head-tail linked list.

fromStrict :: Deque a -> Deque a

Convert strict deque to lazy deque.

toStrict :: Deque a -> Deque a

Convert lazy deque to strict deque.

fromConsAndSnocLists :: [a] -> [a] -> Deque a

\(\mathcal{O}(1)\). Construct from cons and snoc lists.

cons :: a -> Deque a -> Deque a

\(\mathcal{O}(1)\). Add element in the beginning.

snoc :: a -> Deque a -> Deque a

\(\mathcal{O}(1)\). Add element in the ending.

reverse :: Deque a -> Deque a

\(\mathcal{O}(1)\). Reverse the deque.

shiftLeft :: Deque a -> Deque a

\(\mathcal{O}(1)\), occasionally \(\mathcal{O}(n)\). Move the first element to the end.

λ toList . shiftLeft $ fromList [1,2,3]
[2,3,1]

shiftRight :: Deque a -> Deque a

\(\mathcal{O}(1)\), occasionally \(\mathcal{O}(n)\). Move the last element to the beginning.

λ toList . shiftRight $ fromList [1,2,3]
[3,1,2]

filter :: (a -> Bool) -> Deque a -> Deque a

\(\mathcal{O}(n)\). Leave only the elements satisfying the predicate.

take :: Int -> Deque a -> Deque a

\(\mathcal{O}(n)\). Leave only the specified amount of first elements.

drop :: Int -> Deque a -> Deque a

\(\mathcal{O}(n)\). Drop the specified amount of first elements.

takeWhile :: (a -> Bool) -> Deque a -> Deque a

\(\mathcal{O}(n)\). Leave only the first elements satisfying the predicate.

dropWhile :: (a -> Bool) -> Deque a -> Deque a

\(\mathcal{O}(n)\). Drop the first elements satisfying the predicate.

span :: (a -> Bool) -> Deque a -> (Deque a, Deque a)

\(\mathcal{O}(n)\). Perform takeWhile and dropWhile in a single operation.

uncons :: Deque a -> Maybe (a, Deque a)

\(\mathcal{O}(1)\), occasionally \(\mathcal{O}(n)\). Get the first element and deque without it if it's not empty.

unsnoc :: Deque a -> Maybe (a, Deque a)

\(\mathcal{O}(1)\), occasionally \(\mathcal{O}(n)\). Get the last element and deque without it if it's not empty.

null :: Deque a -> Bool

\(\mathcal{O}(1)\). Check whether deque is empty.

head :: Deque a -> Maybe a

\(\mathcal{O}(1)\), occasionally \(\mathcal{O}(n)\). Get the first element if deque is not empty.

last :: Deque a -> Maybe a

\(\mathcal{O}(1)\), occasionally \(\mathcal{O}(n)\). Get the last element if deque is not empty.

tail :: Deque a -> Deque a

\(\mathcal{O}(1)\), occasionally \(\mathcal{O}(n)\). Keep all elements but the first one.

In case of empty deque returns an empty deque.

init :: Deque a -> Deque a

\(\mathcal{O}(1)\), occasionally \(\mathcal{O}(n)\). Keep all elements but the last one.

In case of empty deque returns an empty deque.